- moved content of page where you can edit text from App.tsx to the file EditPage.tsx within the new directory pages - added routing with router from to App.tsx - added LandingPage.tsx to pages directory
